Move `admin inspect` to `support inspect` (#3971)
